1st attempt at making this HEKA PatchMaster version independent
I've created a single function that can read any level of the tree in the HEKA file. Before there were separate functions for each level.
Also, the function should now be able to read files from any version. If the tree level header is smaller than the maximum in the code, it should try to only read up to the length specified in the file itself.

This does need testing!